Logging - added kotlin-logging-jvm dependency to luminositylabs-kotlin-base module - refactored helloworld-1 test module to be located in OSS package and use kotlin logging
