+++ testing/helloworld-1/src/main/java/co/luminositylabs/oss/testing/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Test code.
+ */
+package co.luminositylabs.oss.testing;
