--- testing/helloworld-1/src/main/java/co/luminositylabs/testing/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Test code.
- */
-package co.luminositylabs.testing;
