+++ testing/helloworld-1/src/main/kotlin/co/luminositylabs/oss/testing/KotlinHelloWorld.kt
@@ -1,11 +1,15 @@
-package co.luminositylabs.testing
+package co.luminositylabs.oss.testing
+
+import io.github.oshai.kotlinlogging.KotlinLogging
+
+private val logger = KotlinLogging.logger {}
 
 class KotlinHelloWorld {
     val greetingPrefix = "Hello "
 
     fun sayHello(name: String): String {
         val greeting = "$greetingPrefix$name"
-        println(greeting)
+        logger.debug { greeting }
         return greeting
     }
 }
