+++ testing/helloworld-1/src/main/kotlin/co/luminositylabs/oss/testing/Main.kt
@@ -1,4 +1,4 @@
-package co.luminositylabs.testing
+package co.luminositylabs.oss.testing
 
 fun main(args: Array<String>) {
     println("Hello World!")
