--- testing/helloworld-1/src/test/java/co/luminositylabs/testing/JHelloWorldTest.java
@@ -1,16 +1,0 @@
-package co.luminositylabs.testing;
-
-
-import org.testng.Assert;
-import org.testng.annotations.Test;
-
-
-public class JHelloWorldTest {
-
-    @Test
-    public void test() {
-        final String testName = "Boba[" + getClass().getSimpleName() + "]";
-        Assert.assertEquals("Hello " + testName, new JavaHelloWorld().sayHello(testName));
-    }
-
-}
