+++ testing/helloworld-1/src/test/kotlin/co/luminositylabs/oss/testing/KHelloWorldIT.kt
@@ -1,0 +1,16 @@
+package co.luminositylabs.oss.testing
+
+import io.github.oshai.kotlinlogging.KotlinLogging
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+private val logger = KotlinLogging.logger {}
+
+class KHelloWorldIT {
+    @Test
+    fun `Integration Test`() {
+        logger.trace { "integration test" }
+        val testName = "Boba"
+        assertEquals("Hello $testName", KotlinHelloWorld().sayHello(testName))
+    }
+}
