+++ testing/helloworld-1/src/test/kotlin/co/luminositylabs/oss/testing/KHelloWorldTest.kt
@@ -1,0 +1,16 @@
+package co.luminositylabs.oss.testing
+
+import io.github.oshai.kotlinlogging.KotlinLogging
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+private val logger = KotlinLogging.logger {}
+
+class KHelloWorldTest {
+    @Test
+    fun `Unit Test`() {
+        logger.trace { "unit test" }
+        val testName = "Boba"
+        assertEquals("Hello $testName", KotlinHelloWorld().sayHello(testName))
+    }
+}
