+++ testing/helloworld-1/src/test/resources/logback-test.xml
@@ -1,0 +1,20 @@
+<configuration debug="false">
+
+    <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
+        <layout class="ch.qos.logback.classic.PatternLayout">
+            <Pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} [%M] - %msg%n</Pattern>
+        </layout>
+    </appender>
+
+    <root level="debug">
+        <appender-ref ref="STDOUT" />
+    </root>
+
+    <logger name="co.luminositylabs.oss.testing.JavaHelloWorld" level="TRACE"/>
+    <logger name="co.luminositylabs.oss.testing.KotlinHelloWorld" level="TRACE"/>
+    <logger name="co.luminositylabs.oss.testing.JHelloWorldIT" level="TRACE"/>
+    <logger name="co.luminositylabs.oss.testing.JHelloWorldTest" level="TRACE"/>
+    <logger name="co.luminositylabs.oss.testing.KHelloWorldIT" level="TRACE"/>
+    <logger name="co.luminositylabs.oss.testing.KHelloWorldTest" level="TRACE"/>
+
+</configuration>
